% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_ef.R
\name{age_ef}
\alias{age_ef}
\alias{exp_age_ef}
\title{Age Effect}
\usage{
exp_age_ef(n_age, min_rate = 0.00025, rr = 10)
}
\arguments{
\item{n_age}{number of age groups}

\item{min_rate}{rate in the first age group}

\item{rr}{ratio of the rate in the last age group to that in the first age group}
}
\value{
a function that takes integer indices of age groups and returns the logarithmic rates
}
\description{
Age Effect
}
