% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apc.R
\name{apc_dmat}
\alias{apc_dmat}
\title{Design Matrix of Age-Period-Cohort Model}
\usage{
apc_dmat(A, P, C, degree = 1)
}
\arguments{
\item{A}{Age}

\item{P}{Period}

\item{C}{Cohort}

\item{degree}{1 for linear APC model; 2 for quadratic APC model}
}
\value{
A matrix object
}
\description{
Generate the APC design matrix using the \code{designmatrix()} function from \href{https://analysistools.cancer.gov/apc/}{APC Analysis Web Tool}.
    The source code can be found in its \href{https://github.com/CBIIT/nci-webtools-dceg-age-period-cohort}{Github repository}.
    The quadratic design matrix is implemented by Chernyavskiy et al. (2019).
}
\references{
Philip S. Rosenberg, David P. Check, William F. Anderson.
    A Web Tool for Age–Period–Cohort Analysis of Cancer Incidence and Mortality Rates.
    Cancer Epidemiology, Biomarkers & Prevention (2014); 23(11):2296–2302.

Chernyavskiy P, Little MP, Rosenberg PS.
    A unified approach for assessing heterogeneity in age–period–cohort model parameters using random effects.
    Statistical Methods in Medical Research (2019); 28(1):20-34.
}
