% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-apc_example.R
\docType{data}
\name{apc_example}
\alias{apc_example}
\alias{apc_example1}
\alias{apc_example2}
\alias{apc_example3}
\title{Simulated age-period data by administrative units}
\format{
Each is a data frame with 2,565 rows and 5 columns.

\describe{
  \item{TOWNCODE}{Country name}
  \item{AGE}{Age group}
  \item{PERIOD}{Calendar period}
  \item{POP}{Population size}
  \item{CASE}{Number of cases}
}
}
\usage{
apc_example1

apc_example2

apc_example3
}
\description{
Simulated age-period data by administrative units, including
  57 administrative units, 5 period groups, and 9 age groups.
}
\keyword{datasets}
