% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{buffer}
\alias{buffer}
\title{Create Circular Buffers}
\usage{
buffer(data, ref, buffvar, timevar, sdist, tdist)
}
\arguments{
\item{data}{data for which buffers need to be added}

\item{ref}{data used as a reference to generate buffers}

\item{buffvar}{variable for which buffers are created}

\item{timevar}{time variable}

\item{sdist, tdist}{buffer sizes for space (in meters) and time}
}
\value{
the design matrix for buffers
}
\description{
Create Circular Buffers
}
