% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgm.R
\name{get_nugget}
\alias{get_nugget}
\alias{get_sill}
\title{Extract Nugget and Partial Sill from a Variogram Model}
\usage{
get_nugget(model)

get_sill(model)
}
\arguments{
\item{model}{object of class \code{variogramModel} or \code{StVariogramModel}}
}
\description{
Extract Nugget and Partial Sill from a Variogram Model
}
