% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krige_sp.R
\name{krige_sp}
\alias{krige_sp}
\title{Spatial Kriging (Ordinary or Universal)}
\usage{
krige_sp(formula, data, model, errorVar, min.nug = 0)
}
\arguments{
\item{formula}{formula that defines the response variable as a linear model of covariates}

\item{data}{object of class \code{Spatial} or \code{sf} that contains the response variable, covariates, and coordinates}

\item{model}{variogram model}

\item{errorVar}{variances of measurement errors}

\item{min.nug}{minimum nugget value to avoid singular covariance matrix}
}
\description{
Spatial Kriging (Ordinary or Universal)
}
\references{
Hsu CC, Tsai DR, Su SY, Jhuang JR, Chiang CJ, Yang YW, Lee WC.
A Stabilized Kriging Method for Mapping Disease Rates.
Journal of Epidemiology (2023); 33(4):201-208.
}
