% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krige_st.R
\name{krige_st}
\alias{krige_st}
\title{Spatiotemporal Kriging (Ordinary or Universal)}
\usage{
krige_st(formula, data, model, errorVar, min.nug = 0)
}
\arguments{
\item{formula}{formula that defines the response variable as a linear model of covariates}

\item{data}{object of class \code{ST*DF} that contains the response variable, covariates, and coordinates}

\item{model}{spatio-temporal variogram model}

\item{errorVar}{variances of measurement errors}

\item{min.nug}{minimum nugget value to avoid singular covariance matrix}
}
\description{
Spatiotemporal Kriging (Ordinary or Universal)
}
\references{
Tsai DR, Jhuang JR, Su SY, Chiang CJ, Yang YW, Lee WC.
A stabilized spatiotemporal kriging method for disease mapping and application to male oral cancer and female breast cancer in Taiwan.
BMC Medical Research Methodology (2022); 22(1):270.
}
