% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krige_sp.R
\name{predict.spkrige}
\alias{predict.spkrige}
\title{Prediction for Spatial Kriging}
\usage{
\method{predict}{spkrige}(object, newdata, computeVar = FALSE, ...)
}
\arguments{
\item{object}{fitted spatial Kriging model}

\item{newdata}{object of class \code{Spatial} or \code{sf} with prediction locations; should contain attributes with the covariates (if any)}

\item{computeVar}{logical; compute prediction variances or not}

\item{...}{arguments passed to or from other methods}
}
\description{
Prediction for Spatial Kriging
}
