% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krige_st.R
\name{predict.stkrige}
\alias{predict.stkrige}
\title{Prediction for Spatiotemporal Kriging}
\usage{
\method{predict}{stkrige}(object, newdata, computeVar = FALSE, ...)
}
\arguments{
\item{object}{fitted spatiotemporal Kriging model}

\item{newdata}{object of class \verb{ST*} with prediction locations; should contain attributes with the covariates (if any)}

\item{computeVar}{logical; compute prediction variances or not}

\item{...}{arguments passed to or from other methods}
}
\description{
Prediction for Spatiotemporal Kriging
}
