% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{r2pois}
\alias{r2pois}
\title{Poisson Deviance-based \eqn{R^2} for Count Data}
\usage{
r2pois(y_true, y_pred, na.rm = FALSE)
}
\arguments{
\item{y_true}{true values}

\item{y_pred}{predicted values}

\item{na.rm}{logical; whether NA values should be removed before the computation proceeds}
}
\value{
A numeric value
}
\description{
Calculate the Poisson deviance-based pseudo-\eqn{R^2} for count data.
}
\references{
A. Colin Cameron and Frank A. G. Windmeijer.
R-Squared Measures for Count Data Regression Models With Applications to Health-Care Utilization.
Journal of Business & Economic Statistics (1996); 14(2):209-220.
}
