% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulation}
\alias{simulation}
\alias{st_risk}
\alias{st_risk1}
\alias{st_risk2}
\alias{st_risk3}
\alias{st_risk4}
\title{Spatiotemporal Risk Simulation}
\usage{
st_risk(
  s,
  t,
  age,
  c1,
  c2 = c1,
  c_start,
  c_end,
  bw_stat,
  bw_dyn,
  anis_stat = c(0, 1),
  anis_dyn = c(0, 1),
  rr_stat,
  rr_dyn,
  t_lim,
  t_turn,
  t_par,
  n_arm,
  arm_bw,
  arm_tight,
  arm_eye,
  arm_rot,
  age_fun,
  scenario
)

st_risk1(...)

st_risk2(...)

st_risk3(...)

st_risk4(
  ...,
  n_arm = 3L,
  arm_bw = 0.5,
  arm_tight = 1,
  arm_eye = 0.01,
  arm_rot = 0
)
}
\arguments{
\item{s}{spatial coordinate}

\item{t}{index of time groups (a vector)}

\item{age}{index of age groups (a single value)}

\item{c1}{center of the first static hotspot}

\item{c2}{center of the second static hotspot}

\item{c_start}{starting center of the dynamic hotspot}

\item{c_end}{ending center of the dynamic hotspot}

\item{bw_stat}{bandwidth of the static hotspot(s)}

\item{bw_dyn}{bandwidth of the dynamic hotspot}

\item{anis_stat}{anisotropy parameters of the static hotspot(s); see notes of \code{\link[gstat:vgm]{gstat::vgm()}}}

\item{anis_dyn}{anisotropy parameters of the dynamic hotspot; see notes of \code{\link[gstat:vgm]{gstat::vgm()}}}

\item{rr_stat}{ratio of the highest and lowest disease rates for the static hotspot(s)}

\item{rr_dyn}{ratio of the highest and lowest disease rates for the dynamic hotspot}

\item{t_lim}{time limits}

\item{t_turn}{turning point in time}

\item{t_par}{time parameter}

\item{n_arm}{number of spiral arms}

\item{arm_bw}{bandwidth of spiral arms}

\item{arm_tight}{spiral tightness parameter}

\item{arm_eye}{size of the spiral center as a fraction of \code{bw_dyn}}

\item{arm_rot}{angular velocity of the spinning spiral (in degrees per unit time)}

\item{age_fun}{a function of age that outputs age-specific baseline rates (in log scale)}

\item{scenario}{simulation scenario
\enumerate{
\item convergence–dissipation
\item outward rippling
\item expansion–contraction gaussian
\item expansion–contraction spinning spiral
}}

\item{...}{arguments to be passed to \code{st_risk()}}
}
\value{
a matrix of disease rates in log scale
}
\description{
Spatiotemporal Risk Simulation
}
