% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgm.R
\name{slice_stvgm}
\alias{slice_stvgm}
\title{Slice Plot of Space-Time Variogram Models}
\usage{
slice_stvgm(x, model = NULL, aspect = 1, col = sp::bpy.colors(), ...)
}
\arguments{
\item{x}{object of class \code{StVariogram}}

\item{model}{object of class \code{StVariogramModel}}

\item{aspect}{slicing aspect}

\item{col}{line colors}

\item{...}{arguments passed to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}}
}
\description{
Slice Plot of Space-Time Variogram Models
}
