% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{smape}
\alias{smape}
\title{Symmetric Mean Absolute Percentage Error (SMAPE)}
\usage{
smape(y_true, y_pred, na.rm = FALSE)
}
\arguments{
\item{y_true}{true values}

\item{y_pred}{predicted values}

\item{na.rm}{logical; whether NA values should be removed before the computation proceeds}
}
\value{
A numeric value
}
\description{
Calculate the symmetric mean absolute percentage error (SMAPE).
}
