% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{st_rand_surface}
\alias{st_rand_surface}
\title{Spatiotemporal Random Surface Generator}
\usage{
st_rand_surface(
  sgrid,
  tgrid,
  vgrid = c(-1, 1),
  anis_angle = 1:180,
  anis_ratio = 1,
  sbw,
  tbw,
  n = 101,
  seed
)
}
\arguments{
\item{sgrid}{spatial grid}

\item{tgrid}{temporal grid}

\item{vgrid}{grid of values multiplied to the kernels}

\item{anis_angle}{anisotropy angle, the angle of the main axis direction}

\item{anis_ratio}{anisotropy ratio, the ratio of the minor range to the major range}

\item{sbw}{spatial bandwidth}

\item{tbw}{temporal bandwidth}

\item{n}{number of grid points to sample}

\item{seed}{random seed}
}
\value{
a function that takes spatial and temporal coordinates and returns values of the random surface
}
\description{
Spatiotemporal Random Surface Generator
}
