% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stablemap-package.R
\docType{package}
\name{stablemap-package}
\alias{stablemap}
\alias{stablemap-package}
\title{stablemap: Stabilized Disease Mapping}
\description{
This package provides statistical methods for stabilized disease mapping and tools for generating space-time simulation data. The supported methods include: stabilized kriging (Hsu et al., 2023) <\doi{10.2188/jea.JE20210276}>, stabilized spatiotemporal kriging (Tsai et al., 2022) <\doi{10.1186/s12874-022-01749-9}>, age-period-cohort kriging, etc.
}
\author{
\strong{Maintainer}: Dai-Rong Tsai \email{darrentsai1222@gmail.com} (\href{https://orcid.org/0000-0003-1372-4473}{ORCID})

}
\keyword{internal}
