% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{age_ef}
\alias{age_ef}
\alias{exp_age_ef}
\title{Exponential Age Effect}
\usage{
exp_age_ef(age_n, min_rate = 0.00025, rr = 10)
}
\arguments{
\item{age_n}{number of age groups}

\item{min_rate}{rate in the first age group}

\item{rr}{ratio of the rate in the last age group to that in the first age group}
}
\description{
Exponential Age Effect
}
