% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apc.R
\name{apc_dmat}
\alias{apc_dmat}
\title{Design Matrix of Age-Period-Cohort Model}
\usage{
apc_dmat(A, P, C, degree = 1)
}
\arguments{
\item{A}{Age}

\item{P}{Period}

\item{C}{Cohort}

\item{degree}{1 for linear APC model; 2 for quadratic APC model}
}
\value{
A matrix object
}
\description{
Construct the APC design matrix with an age-drift parameterization.
}
