% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krige.R
\name{krige_sp}
\alias{krige_sp}
\title{Spatial Ordinary or Universal Kriging}
\usage{
krige_sp(
  formula,
  data,
  newdata,
  model,
  errorVar,
  computeVar = FALSE,
  epsilon = 1e-07
)
}
\arguments{
\item{formula}{formula that defines the response variable as a linear model of covariates}

\item{data}{object of class `Spatial` or `sf` that contains the response variable, covariates, and coordinates}

\item{newdata}{object of class `Spatial` or `sf` with prediction locations; should contain attributes with the covariates (if any)}

\item{model}{variogram model}

\item{errorVar}{variances of measurement errors}

\item{computeVar}{logical; compute prediction variances or not}

\item{epsilon}{small value added to nugget to avoid singular covariance matrix}
}
\description{
Spatial Ordinary or Universal Kriging
}
\references{
Hsu CC, Tsai DR, Su SY, Jhuang JR, Chiang CJ, Yang YW, Lee WC.
    A Stabilized Kriging Method for Mapping Disease Rates.
    Journal of Epidemiol (2023); 33(4):201-208.
}
