% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulation}
\alias{simulation}
\alias{scenario1}
\alias{scenario2}
\alias{scenario3}
\title{Spatiotemporal Simulation}
\usage{
scenario1(
  s,
  t,
  age,
  c1,
  c2,
  c_start,
  c_end,
  s_bw_static,
  s_bw_dynamic,
  t_n,
  t_turn,
  t_bw,
  age_fun = exp_age_ef
)

scenario2(
  s,
  t,
  age,
  c1,
  c2,
  c_start,
  c_end,
  s_bw_static,
  s_bw_dynamic,
  t_n,
  t_turn,
  t_bw,
  age_fun = exp_age_ef
)

scenario3(
  s,
  t,
  age,
  c1,
  c2,
  c_start,
  c_end,
  s_bw_static,
  s_bw_dynamic,
  t_n,
  t_turn,
  t_bw,
  age_fun = exp_age_ef
)
}
\arguments{
\item{s}{spatial coordinate}

\item{t}{index of time groups}

\item{age}{index of age groups}

\item{c1}{center of the first static hotspot}

\item{c2}{center of the second static hotspot}

\item{c_start}{starting center of the dynamic hotspot}

\item{c_end}{ending center of the dynamic hotspot}

\item{s_bw_static}{bandwidth of static hotspot}

\item{s_bw_dynamic}{bandwidth of dynamic hotspot}

\item{t_n}{number of time groups}

\item{t_turn}{turning time point}

\item{t_bw}{bandwidth in time}

\item{age_fun}{age effect function}
}
\description{
Spatiotemporal Simulation
}
