% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulation}
\alias{simulation}
\alias{st_sim}
\alias{st_sim1}
\alias{st_sim2}
\alias{st_sim3}
\title{Spatiotemporal Simulation}
\usage{
st_sim(
  s,
  t,
  age,
  c1,
  c2,
  c_start,
  c_end,
  s_bw_static,
  s_bw_dynamic,
  rr_static,
  rr_dynamic,
  t_n,
  t_turn,
  t_bw,
  age_fun,
  scenario
)

st_sim1(...)

st_sim2(...)

st_sim3(...)
}
\arguments{
\item{s}{spatial coordinate}

\item{t}{index of time groups (a vector)}

\item{age}{index of age groups (a single value)}

\item{c1}{center of the first static hotspot}

\item{c2}{center of the second static hotspot}

\item{c_start}{starting center of a dynamic hotspot}

\item{c_end}{ending center of a dynamic hotspot}

\item{s_bw_static}{bandwidth of a static hotspot}

\item{s_bw_dynamic}{bandwidth of a dynamic hotspot}

\item{rr_static}{ratio of the highest and lowest disease rates for a static hotspot}

\item{rr_dynamic}{ratio of the highest and lowest disease rates for a dynamic hotspot}

\item{t_n}{number of time groups}

\item{t_turn}{turning time point}

\item{t_bw}{bandwidth in time}

\item{age_fun}{a function of age that outputs age-specific baseline rates}

\item{scenario}{simulation scenario
\itemize{
  \item 1: convergence–dissipation
  \item 2: outward rippling
  \item 3: expansion–contraction
}}

\item{...}{arguments to be passed to \code{st_sim()}}
}
\value{
A matrix of logarithmic disease rates
}
\description{
Spatiotemporal Simulation
}
