% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{smape}
\alias{smape}
\title{Symmetric Mean Absolute Percentage Error}
\usage{
smape(x, y)
}
\arguments{
\item{x}{predicted values}

\item{y}{true values}
}
\value{
A numeric value
}
\description{
Symmetric Mean Absolute Percentage Error
}
